moving filter next to search
